feature: make change password button and etc.
